Remove `inserted_at` from migration metadata
Not sure about this, but for now let's keep it off. It may be helpful at
some very specific situation (specially during debugging) BUT it causes
discrepancies when `dump`ing the structure of a database.

I think having a reliable dumping strategy (so that the structure of the
database(s) can be kept in version control -- easier code review) trumps
an hypothetical debugging improvement.