Refactor Boot shard migration to find any shards
Previously, we were looking for shards sequentially, and we'd stop
looking as soon as the next element in the sequence could not be found.

It is very possible that one particular user shard (say, shard_id=42)
had to be deleted for any reason (maybe user was deleted/banned). The
previous logic would not migrate any shard > 42.

This current implementation eagerly finds all shards, which is a better
approach BUT I'm not so sure how it would cope with handling tens of
thousands of shards being migrated -- can it fit everything in memory
and is performance acceptable? That, however, is out of scope for what
I'm currently working at. I'll look into this in the future.