Add support for enqueuing requests when there are no available Repos
Each shard can have at most 1 write Repo and at most 2 read Repos.

The "single-write Repo" is a fundamental constraint of SQLite and will
never change. The "at most 2 read Repos" bit is arbitrary from my part
and can be tweaked in the future (for example, to allow a configurable
number of read Repos) -- it's fine to have multiple concurrent readers.

Previously, when the requested connection could not be fetched (because
they were already in use by another process), the Repo.Manager would
simply return a `:busy` atom.

Now, it will block the caller (at `Repo.Manager.fetch_connection/2`) and
return a response once a connection is made available.

Future improvements should be made to implement a timeout timer, so that
the caller does not block indefinitely.