Add support for after_read callbacks
These callbacks are executed after every other field in the schema has
been loaded. It receives two arguments: the loaded field as well as the
entire loaded schema.

Receiving the entire schema is particularly helpful if the callback
needs to have access to some other data from the row in order to execute
its logic.

Note that if multiple fields implement the `after_read` callback,
there's no guarantee of order over which gets executed first, so one
should make sure that in this scenario the order is not strictly
necessary and/or can be coordinated via some external medium (e.g.
process dictionary).