Update Map type to default to `keys: :atom`
This may not be the safest default, but it's the option I use the most
so I guess it makes sense to make it the default.

It is the developer's responsibility to read the docs and figure out
whether there are options/flags better suited for his workload. When
storing maps with potentially arbitrary keys, one should use either
`:string` or `:safe_atom` as value for `:keys`.